(chore): refactor test suite
